refactor: move csv parser
