+++ src/Import/Csv/Csv.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * @license    LGPL-3.0-or-later
  */
 
-namespace Cgoit\ContaoCalendarIcalBundle\Classes;
+namespace Cgoit\ContaoCalendarIcalBundle\Import\Csv;
 
 class Csv
 {
